fixed bugs

+++ database/migrations/2023_03_08_201823_create_companies_table.php
@@ -15,13 +15,13 @@ return new class extends Migration
     {
         Schema::create('companies', function (Blueprint $table) {
             $table->id();
-            
+
             $table->string('name');
             $table->string('director');
             $table->string('address');
             $table->string('email');
             $table->string('website');
-            $table->integer('phone');
+            $table->bigInteger('phone');
 
             $table->timestamps();
         });
